udpate icon, bump dep

+++ SSH/build.ps1
@@ -1,35 +1,51 @@
 param(
-    [switch]$skipBuild
+	[switch]$skipBuild,
+	[switch]$skipBump,
+	[switch]$skipSign
 )
+
+$ErrorActionPreference = 'Stop'
 
 Push-Location $PSScriptRoot
 
-$version = Read-Host -Prompt 'New tag'
+if (-not $skipBump){
+	$version = Read-Host -Prompt 'New tag'
 
-(Get-Content ./app.manifest -Raw) -replace `
-	   '<assemblyIdentity version="[\d\.]+"', `
-	   "<assemblyIdentity version=`"$version.0`"" `
-| Out-File .\app.manifest -NoNewline
-(Get-Content ./Package.appxmanifest -Raw) -replace `
-	   '    Version="[\d\.]+" />', `
-	   "    Version=`"$version.0`" />" `
-| Out-File ./Package.appxmanifest -NoNewline
-(Get-Content ./SSH.csproj -Raw) -replace `
-    '<Version>[\d\.]+</Version>', `
-    "<Version>$version</Version>" `
-| Out-File ./SSH.csproj -NoNewline
-
-if (-not $skipBuild) {
-    dotnet build -c Release -p:GenerateAppxPackageOnBuild=true -p:Platform=x64
-    dotnet build -c Release -p:GenerateAppxPackageOnBuild=true -p:Platform=arm64
-    Remove-Item ./out/* -ea Ignore
-    mkdir ./out -ea ig
-    (Get-ChildItem -r ./bin/**/*.msix -Exclude Microsoft.WindowsAppRuntime.*.msix).FullName | ForEach-Object { Copy-Item $_ ./out/. -Force }
+	(Get-Content ./Package.appxmanifest -Raw) -replace `
+		'(Identity[^>]+?)Version="[\d\.]+"', `
+		"`$1Version=`"$version.0`"" `
+	| Out-File ./Package.appxmanifest -NoNewline
 }
 
-git add ..
-git commit -m 'bump'
-git tag "v$version"
+if (-not $skipBuild) {
+	Remove-Item ./AppPackages/* -r -fo -ea ig
+	dotnet build --configuration Release -p:Platform=x64 -p:AppxPackageDir="AppPackages\x64\"
+	dotnet build --configuration Release -p:Platform=arm64 -p:AppxPackageDir="AppPackages\ARM64\"
+	mkdir out -ea ig
+	Remove-Item ./out/* -r -fo -ea ig
+	$msix = Get-ChildItem ./AppPackages/ -Recurse -Filter *.msix
+	Copy-Item $msix ./out/. -Force
+
+	Push-Location
+	Import-Module 'C:\Program Files\Microsoft Visual Studio\18\Community\Common7\Tools\Microsoft.VisualStudio.DevShell.dll'
+	Enter-VsDevShell e87fa5ce
+	Pop-Location
+	$filename = $msix[0].Name
+	$firstIndex = $filename.IndexOf('_')
+	$secondIndex = $filename.IndexOf('_', $firstIndex + 1)
+	$bundle = "$($filename.Substring(0, $secondIndex)).msixbundle"
+	makeappx bundle /v /d ./out/ /p $bundle
+	Move-Item $bundle ./out
+}
+
+if (-not $skipSign) {
+	winapp sign ./out/$bundle E:/cert/cert.pfx
+}
+
+if (-not $skipBump){
+	git add ..
+	git commit -m 'bump'
+	git tag "v$version"
+}
 
 Pop-Location
-
